+++ tests/test_models.py
@@ -192,13 +192,13 @@ class TestGaussian(FitPredictParent):
     def test_2_factors_saturated(self, crossed_data):
         model = bmb.Model("Y ~ threecats*fourcats", crossed_data)
         idata = self.fit(model)
-        assert list(idata.posterior.data_vars) == [
+        assert set(idata.posterior.data_vars) == {
             "Intercept",
             "threecats",
             "fourcats",
             "threecats:fourcats",
             "Y_sigma",
-        ]
+        }
         assert list(idata.posterior["threecats_dim"].values) == ["b", "c"]
         assert list(idata.posterior["fourcats_dim"].values) == ["b", "c", "d"]
         assert list(idata.posterior["threecats:fourcats_dim"].values) == [
@@ -214,12 +214,12 @@ class TestGaussian(FitPredictParent):
     def test_2_factors_no_intercept(self, crossed_data):
         model = bmb.Model("Y ~ 0 + threecats*fourcats", crossed_data)
         idata = self.fit(model)
-        assert list(idata.posterior.data_vars) == [
+        assert set(idata.posterior.data_vars) == {
             "threecats",
             "fourcats",
             "threecats:fourcats",
             "Y_sigma",
-        ]
+        }
         assert list(idata.posterior["threecats_dim"].values) == ["a", "b", "c"]
         assert list(idata.posterior["fourcats_dim"].values) == ["b", "c", "d"]
         assert list(idata.posterior["threecats:fourcats_dim"].values) == [
@@ -235,7 +235,7 @@ class TestGaussian(FitPredictParent):
     def test_2_factors_cell_means(self, crossed_data):
         model = bmb.Model("Y ~ 0 + threecats:fourcats", crossed_data)
         idata = self.fit(model)
-        assert list(idata.posterior.data_vars) == ["threecats:fourcats", "Y_sigma"]
+        assert set(idata.posterior.data_vars) == {"threecats:fourcats", "Y_sigma"}
         assert list(idata.posterior["threecats:fourcats_dim"].values) == [
             "a, a",
             "a, b",
@@ -255,7 +255,7 @@ class TestGaussian(FitPredictParent):
     def test_cell_means_with_covariate(self, crossed_data):
         model = bmb.Model("Y ~ 0 + threecats + continuous", crossed_data)
         idata = self.fit(model)
-        assert list(idata.posterior.data_vars) == ["threecats", "continuous", "Y_sigma"]
+        assert set(idata.posterior.data_vars) == {"threecats", "continuous", "Y_sigma"}
         assert list(idata.posterior["threecats_dim"].values) == ["a", "b", "c"]
         self.predict_oos(model, idata)
 
@@ -477,7 +477,7 @@ class TestGaussian(FitPredictParent):
         idata = self.fit(model)
         self.predict_oos(model, idata)
 
-        assert list(idata.posterior.data_vars) == [
+        assert set(idata.posterior.data_vars) == {
             "Intercept",
             "continuous",
             "Y_sigma",
@@ -485,7 +485,7 @@ class TestGaussian(FitPredictParent):
             "threecats:fourcats|site_sigma",
             "1|site",
             "threecats:fourcats|site",
-        ]
+        }
         assert list(idata.posterior["threecats:fourcats|site"].coords) == [
             "chain",
             "draw",
@@ -638,19 +638,29 @@ class TestBinomial(FitPredictParent):
         self.assert_mean_range(model, idata)
         assert (idata.posterior_predictive["prop(y, n)"].to_numpy() <= y_reshaped).all()
 
+        # Test log-likelihood computation
+        model.compute_log_likelihood(idata)
+        idata_2 = model.compute_log_likelihood(idata, data=beetle_data, inplace=False)
+        assert (idata_2.log_likelihood["prop(y, n)"] == idata.log_likelihood["prop(y, n)"]).all().item()
+
     def test_binomial_regression_constant(self, beetle_data):
         # Uses a constant instead of variable in data frame
-        model = bmb.Model("prop(y, 62) ~ x", beetle_data, family="binomial")
+        model = bmb.Model("p(y, 62) ~ x", beetle_data, family="binomial")
         idata = self.fit(model)
         model.predict(idata, kind="pps")
         self.assert_mean_range(model, idata)
-        assert (idata.posterior_predictive["prop(y, 62)"].to_numpy() <= 62).all()
-        assert (0 <= idata.posterior_predictive["prop(y, 62)"].to_numpy()).all()
+        assert (idata.posterior_predictive["p(y, 62)"].to_numpy() <= 62).all()
+        assert (0 <= idata.posterior_predictive["p(y, 62)"].to_numpy()).all()
 
         # Out of sample prediction
         idata = self.predict_oos(model, idata)
         self.assert_mean_range(model, idata)
-        assert (idata.posterior_predictive["prop(y, 62)"].to_numpy() <= 62).all()
+        assert (idata.posterior_predictive["p(y, 62)"].to_numpy() <= 62).all()
+
+        # Test log-likelihood computation
+        model.compute_log_likelihood(idata)
+        idata_2 = model.compute_log_likelihood(idata, data=beetle_data, inplace=False)
+        assert (idata_2.log_likelihood["p(y, 62)"] == idata.log_likelihood["p(y, 62)"]).all().item()
 
 
 class TestPoisson(FitPredictParent):
@@ -787,6 +797,11 @@ class TestGamma(FitPredictParent):
         assert set(idata.posterior.data_vars) == {"Intercept", "y2", "y3", "n1", "cat4", "o_alpha"}
         idata = self.predict_oos(model, idata)
         assert (idata.posterior_predictive["o"] > 0).all().item()
+
+        # Compute log likelihood
+        model.compute_log_likelihood(idata)
+        idata_2 = model.compute_log_likelihood(idata, data=data_n100, inplace=False)
+        assert (idata.log_likelihood["o"] == idata_2.log_likelihood["o"]).all().item()
 
     def test_gamma_regression_categoric(self, data_n100):
         data_n100["o"] = np.exp(data_n100["y1"])
@@ -1178,6 +1193,12 @@ class TestMultinomial(FitPredictParent):
         idata = self.fit(model, random_seed=121195)
         idata = self.predict_oos(model, idata, data=model.data)
         self.assert_posterior_predictive(model, idata)
+
+        # Log likelihood computation
+        model.compute_log_likelihood(idata)
+        idata_2 = model.compute_log_likelihood(idata, data=multinomial_data, inplace=False)
+        name = "c(y1, y2, y3, y4)"
+        assert (idata.log_likelihood[name] == idata_2.log_likelihood[name]).all().item()
 
     def test_categorical_predictors(self, multinomial_data):
         multinomial_data["treat"] = multinomial_data["treat"].replace({-0.5: "A", 0.5: "B"})