+++ CHANGELOG.md
@@ -8,10 +8,14 @@
 * Interpet submodule now outputs informative messages when computing default values (#745) 
 * Bambi supports weighted responses (#761)
 * Bambi supports constrained responses (#764)
+* Implement `compute_log_likelihood()` method to compute the log likelihood on a model (#769)
 
 ### Maintenance and fixes
 
 * Fix bug in predictions with models using HSGP (#780)
+* Fix `get_model_covariates()` utility function (#801)
+* Upgrade PyMC dependency to >= 5.13 (#803)
+* Use `pm.compute_deterministics()` to compute deterministics when bayeux based samplers are used (#803)
 
 ### Documentation